bunch of stuff - mainly adding timing and getting rid of xpath.
git-svn-id: https://xcmetadataservicestoolkit.googlecode.com/svn/branches/bens_perma_branch@1129 72d1a826-0e72-11de-b5bf-434079a3519f

+++ mst-common/src/java/xc/mst/harvester/HarvestManager.java
@@ -126,7 +126,7 @@ public class HarvestManager extends BaseManager implements WorkDelegate {
 		return recordsProcessed;
 	}
 
-	public int getTotalRecords() {
+	public long getTotalRecords() {
 		return totalRecords;
 	}
